Poprawione bledy zauwazone przez prowadzacego

+++ lab3/Generator2.cpp
@@ -2,7 +2,7 @@
 
 // ---------------------------------------------------------
 // ---------------------------------------------------------
-Generator2::Generator2(int iPoczatek, int iKoniec, unsigned long long seed, unsigned long long mnoznik, unsigned long long add, unsigned long long modul, string strNazwaGeneratora)
+Generator2::Generator2(int iPoczatek, int iKoniec, unsigned long seed, unsigned long mnoznik, unsigned long add, unsigned long modul, string strNazwaGeneratora)
 {
     this->setPoczatek(iPoczatek);
     this->setKoniec(iKoniec);
@@ -21,10 +21,10 @@ Generator2::~Generator2()
 }
 
 // ---------------------------------------------------------
-unsigned long long Generator2::LCG(unsigned long long mnoznik, unsigned long long  seed, unsigned long long modul)
+unsigned long Generator2::LCG(unsigned long mnoznik, unsigned long  seed, unsigned long modul)
 {
 
-    unsigned long long b, n, r;
+    unsigned long b, n, r;
  
     r = 0;
     n = 1;
